Remove support for centos 6.
Graphite 1.x requires at least python 2.7, it is available with SCL but
it's a lot of work to adapt all the init-scripts and Ansible tasks to
work with SCL. I will remove support considering that I guess that there
are not that many people using centos 6 that like to use this role.

If you are using centos 6 and like to install graphite with this role,
use the 2.0 release of this role. It works well and installs graphite
0.9.15.